refactored views into folders
