+++ src/components/Splash.vue
