+++ src/views/public/CaseView.vue
